requirements.txt
langgraph
langchain
langchain-community
langchain-ollama
tree-sitter
tree-sitter-python
tree-sitter-javascript
faiss-cpu
rich
pyyaml
pydantic